+++ src/arch_rogue/input.py
@@ -1315,7 +1315,10 @@ class InputMixin:
             return True
         if cmd == Command.ABILITY_3:
             self._sync_controller_action_aim()
-            self.player_cast_nova()
+            if self.player.class_name == "Acolyte":
+                self.player_cast_spirit_call()
+            else:
+                self.player_cast_nova()
             return True
         if cmd == Command.ABILITY_4:
             self._sync_controller_action_aim()